AppWarnings: Remove old app target SDK dialog
If an app is old, users should already know that, and there's usually no
point in warning them about it because they would already be using a
newer version if one existed. Sometimes, using an old app is necessary
for one reason or another, so remove this annoyance and let the user use
their old app in peace.

Change-Id: Ieb40d529780d52a54ad314beff10c72f5de92eff
Signed-off-by: Richard Raya <rdxzv.dev@gmail.com>