Do not kill process on removing task if it has a started service
Upon removing the task, if the app is running in the background
process state, it'll be killed. We've exempted if it's running
a broadcast receiver, and here we're exempting the apps with
started service, so the app will get a chance to run its
Service#onTaskRemoved(). Since the app is running in the background
state, its background service will be stopped automatically after
60s, so the process will be killed eventually.

Bug: 310724715
Test: Manual - Launch the media session demo, swipe it from recents.
Change-Id: Ie6c17a7b77a6189b8eac1d35d49c862ea46c00b9
Signed-off-by: Richard Raya <rdxzv.dev@gmail.com>